requirements.txt
fastapi
uvicorn
numpy
pillow
pillow-heif
insightface
onnxruntime
python-multipart
